Update OpenCoverParser.cs
Code to update data from sl to el in SequencePoint is causing the Report to count anonymous delegates and lambdas as fully covered even if they have not been exectured.  It also overcounts single code lines that span multiple lines, effecting the coverage percentage.  This code will break the coloring of subsequent lines in the actual case of multiple lines for formatting, but fixes the count.